+++ src/types/Theme.ts
@@ -325,6 +325,25 @@ export interface Theme {
   outlineTextInputLabelColor: string
   outlineTextInputLabelColorFocused: string
 
+  // Text Input
+  textInputTextColor: string
+  textInputTextColorDisabled: string
+  textInputTextColorFocused: string
+  textInputBackgroundColor: string
+  textInputBackgroundColorDisabled: string
+  textInputBackgroundColorFocused: string
+  textInputBorderWidth: number
+  textInputBorderColor: string
+  textInputBorderColorDisabled: string
+  textInputBorderColorFocused: string
+  textInputBorderRadius: number
+  textInputIconColor: string
+  textInputIconColorDisabled: string
+  textInputIconColorFocused: string
+  textInputPlaceholderColor: string
+  textInputPlaceholderColorDisabled: string
+  textInputPlaceholderColorFocused: string
+
   // Dropdown colors:
   dropdownWarning: string
   dropdownError: string